// src/app/models/employee.model.ts
export interface Employee {
  employeeId: number;
  name: string;
  hireDate: Date;
  jobTitle: string;
  email: string;
  hourlyWage: number;
  fica: number;
  lastRaiseDate: Date;
  insurance: number;
  vacationHoliday: number;
  carFuel: number;
  markup: number;
  rate: number;
  burdenedRate: number;
  burdenedOtRate: number;
  notes: string;
  lastVacationDate: Date;
  vacationDaysUsed: number;
}